added a parser for levels and made start with level object

+++ core/src/main/java/sem/group15/bubblebobble/core/objects/FloorObject.java
@@ -11,10 +11,10 @@ import sem.group15.bubblebobble.core.BubbleBobble;
  */
 public class FloorObject extends ImmutableObject {
 
-    public FloorObject(int locationX, int locationY) {
+    public FloorObject(float locationX, float locationY) {
         super(new Rectangle(locationX, locationY, BubbleBobble.SPRITE_SIZE, BubbleBobble.SPRITE_SIZE), new Texture(Gdx.files.internal("sprite_wall_brick.png")));
     }
-    public FloorObject(int locationX, int locationY, Texture texture) {
+    public FloorObject(float locationX, float locationY, Texture texture) {
         super(
                 new Rectangle(locationX, locationY, BubbleBobble.SPRITE_SIZE, BubbleBobble.SPRITE_SIZE),
                 texture
